feat: independent worker loaders
