+++ modules/las/src/las-loader.ts
@@ -13,24 +13,6 @@ export type LASLoaderOptions = LoaderOptions & {
   };
 };
 
-/**
- * Type for header of the .las file
- */
-export type LASHeader = {
-  pointsOffset: number;
-  pointsFormatId: number;
-  pointsStructSize: number;
-  pointsCount: number;
-  scale: [number, number, number];
-  offset: [number, number, number];
-  maxs?: number[];
-  mins?: number[];
-  totalToRead: number;
-  totalRead: number;
-  versionAsString?: string;
-  isCompressed?: boolean;
-};
-
 const DEFAULT_LAS_OPTIONS: LASLoaderOptions = {
   las: {
     fp64: false,
