+++ modules/las/src/lib/laslaz-decoder.ts
@@ -4,7 +4,7 @@
   MIT License
 */
 // laslaz.js - treat as compiled code
-import type {LASHeader} from '../las-loader';
+import type {LASHeader} from './las-types';
 import getModule from '../libs/laz-perf';
 
 let Module: any = null;
