+++ modules/pcd/src/pcd-loader.ts
@@ -1,5 +1,4 @@
-import type {Loader, LoaderWithParser} from '@loaders.gl/loader-utils';
-import parsePCDSync from './lib/parse-pcd';
+import type {Loader} from '@loaders.gl/loader-utils';
 
 // __VERSION__ is injected by babel-plugin-version-inline
 // @ts-ignore TS2304: Cannot find name '__VERSION__'.
@@ -8,7 +7,7 @@ const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
 /**
  * Worker loader for PCD - Point Cloud Data
  */
-export const PCDWorkerLoader: Loader = {
+export const PCDLoader = {
   name: 'PCD (Point Cloud Data)',
   id: 'pcd',
   module: 'pcd',
@@ -21,11 +20,4 @@ export const PCDWorkerLoader: Loader = {
   }
 };
 
-/**
- * Loader for PCD - Point Cloud Data
- */
-export const PCDLoader: LoaderWithParser = {
-  ...PCDWorkerLoader,
-  parse: async (arrayBuffer) => parsePCDSync(arrayBuffer),
-  parseSync: parsePCDSync
-};
+export const _typecheckPCDLoader: Loader = PCDLoader;
