+++ modules/ply/src/ply-loader.ts
@@ -1,19 +1,16 @@
-// description: 'PLY - Polygon File Format (aka Stanford Triangle Format)'
-// links: ['http://paulbourke.net/dataformats/ply/',
-// 'https://en.wikipedia.org/wiki/PLY_(file_format)']
-
-import type {Loader, LoaderWithParser} from '@loaders.gl/loader-utils';
-import parsePLY from './lib/parse-ply';
-import parsePLYInBatches from './lib/parse-ply-in-batches';
+// PLY Loader
+import type {Loader} from '@loaders.gl/loader-utils';
 
 // __VERSION__ is injected by babel-plugin-version-inline
 // @ts-ignore TS2304: Cannot find name '__VERSION__'.
 const VERSION = typeof __VERSION__ !== 'undefined' ? __VERSION__ : 'latest';
 
 /**
- * Worker loader for PLY - Polygon File Format
+ * Worker loader for PLY - Polygon File Format (aka Stanford Triangle Format)'
+ * links: ['http://paulbourke.net/dataformats/ply/',
+ * 'https://en.wikipedia.org/wiki/PLY_(file_format)']
  */
-export const PLYWorkerLoader: Loader = {
+export const PLYLoader = {
   name: 'PLY',
   id: 'ply',
   module: 'ply',
@@ -29,14 +26,4 @@ export const PLYWorkerLoader: Loader = {
   }
 };
 
-/**
- * Loader for PLY - Polygon File Format
- */
-export const PLYLoader: LoaderWithParser = {
-  ...PLYWorkerLoader,
-  // Note: parsePLY supports both text and binary
-  parse: async (arrayBuffer, options) => parsePLY(arrayBuffer, options), // TODO - this may not detect text correctly?
-  parseTextSync: parsePLY,
-  parseSync: parsePLY,
-  parseInBatches: parsePLYInBatches
-};
+export const _typecheckPLYLoader: Loader = PLYLoader;
