+++ modules/terrain/src/quantized-mesh-loader.ts
@@ -1,11 +1,10 @@
-import type {Loader, LoaderWithParser} from '@loaders.gl/loader-utils';
+import type {Loader} from '@loaders.gl/loader-utils';
 import {VERSION} from './lib/utils/version';
-import parseQuantizedMesh from './lib/parse-quantized-mesh';
 
 /**
  * Worker loader for quantized meshes
  */
-export const QuantizedMeshWorkerLoader: Loader = {
+export const QuantizedMeshLoader = {
   name: 'Quantized Mesh',
   id: 'quantized-mesh',
   module: 'terrain',
@@ -20,11 +19,4 @@ export const QuantizedMeshWorkerLoader: Loader = {
   }
 };
 
-/**
- * Loader for quantized meshes
- */
-export const QuantizedMeshLoader: LoaderWithParser = {
-  ...QuantizedMeshWorkerLoader,
-  parseSync: parseQuantizedMesh,
-  parse: async (arrayBuffer, options) => parseQuantizedMesh(arrayBuffer, options)
-};
+export const _typecheckQuantizedMeshLoader: Loader = QuantizedMeshLoader;
